chore: rename statistic-test folder to statistical-analysis
