+++ src/modules/statistical-analysis/components/others/contingency-table.tsx
