+++ src/modules/statistical-analysis/components/plot-config.tsx
