+++ src/modules/statistical-analysis/components/regression/components.tsx
