+++ src/modules/statistical-analysis/components/regression/data.tsx
