+++ src/modules/statistical-analysis/components/regression/linear.tsx
