+++ src/modules/statistical-analysis/components/regression/multinomial-logistic.tsx
