+++ src/modules/statistical-analysis/components/regression/ordinal.tsx
