+++ src/modules/statistical-analysis/components/regression/types.tsx
