+++ src/modules/statistical-analysis/components/regression/utils.ts
