+++ src/modules/statistical-analysis/components/statistic-test/binary-test-contingency-table.tsx
