+++ src/modules/statistical-analysis/components/statistic-test/common.tsx
