+++ src/modules/statistical-analysis/components/statistic-test/pairwise.tsx
