+++ src/modules/statistical-analysis/components/statistic-test/results.tsx
