+++ src/modules/statistical-analysis/configuration/binary-statistic-test.tsx
