+++ src/modules/statistical-analysis/configuration/contingency-table.tsx
