+++ src/modules/statistical-analysis/configuration/logistic-regression.tsx
