+++ src/modules/statistical-analysis/configuration/regression-common.tsx
