+++ src/modules/statistical-analysis/configuration/regression.tsx
