+++ src/modules/statistical-analysis/configuration/select-statistic-test-method.tsx
