+++ src/modules/statistical-analysis/configuration/statistic-test.tsx
