+++ src/modules/statistical-analysis/configuration/utils.tsx
