+++ src/modules/statistical-analysis/data-provider/binary.tsx
