+++ src/modules/statistical-analysis/data-provider/contingency-table.tsx
