+++ src/modules/statistical-analysis/data-provider/regression.tsx
