+++ src/modules/statistical-analysis/data-provider/statistic-test.tsx
