+++ src/modules/statistical-analysis/dictionary.ts
