+++ src/modules/statistical-analysis/statistic-test-config.tsx
