+++ src/modules/visualization/components/configuration/index.tsx
@@ -9,7 +9,7 @@ import {
 import { Faders } from '@phosphor-icons/react';
 import React from 'react';
 export * from './categorical';
-export * from '../../../statistic-test/components/plot-config';
+export * from '../../../statistical-analysis/components/plot-config';
 
 export function PlotInlineConfiguration(props: React.PropsWithChildren) {
   const [opened, { toggle }] = useDisclosure();
