+++ src/modules/visualization/configuration/test-distribution.tsx
@@ -13,7 +13,7 @@ import {
   StatisticMethodSelectField,
   EffectSizeSelectField,
   GroupStatisticMethodSelectField,
-} from '@/modules/statistic-test/configuration/select-statistic-test-method';
+} from '@/modules/statistical-analysis/configuration/select-statistic-test-method';
 import { ProjectContext } from '@/modules/project/context';
 import { ProjectColumnSelectField } from '@/modules/project/select-column-input';
 import { Alert, Group } from '@mantine/core';
